ManipulationCanvas
Delete ManipulationCanvasBean is PanelScript logic file.
Create folder "beans".
Create "ManipulationCanvasBean.cs".
