+++ BallLdyProject/Assets/Scripts/Beans.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: b2382e869e80e6b47a044fc431c15c3b
+folderAsset: yes
+timeCreated: 1601345953
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
