+++ BallLdyProject/Assets/Scripts/Beans/ManipulationCanvasBean.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 1e5cffda6176d8e49ad03b6a99bda512
+timeCreated: 1601345979
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
